Extract reusable helper methods into a shared scope so they can be reused by other FeedStore implementation tests
